backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=1.54.0
httpx>=0.27.0
python-dotenv==1.0.0
chromadb==0.5.23
